refactored some part of bert
