+++ examples/bert/main.py
@@ -7,7 +7,7 @@ Example:
 
 """
 import bittensor
-from bittensor.synapses.bert import BertSynapseConfig, BertMLMSynapse
+from bittensor.synapses.bert import BertSynapseConfig, BertMLMSynapse, mlm_batch
 
 import argparse
 from datasets import load_dataset
@@ -16,32 +16,6 @@ import random
 import torch
 import transformers
 from transformers import DataCollatorForLanguageModeling
-
-def mlm_batch(data, batch_size, tokenizer, collator):
-    """ Returns a random batch from text dataset with 50 percent NSP.
-
-        Args:
-            data: (List[dict{'text': str}]): Dataset of text inputs.
-            batch_size: size of batch to create.
-        
-        Returns:
-            tensor_batch torch.Tensor (batch_size, sequence_length): List of tokenized sentences.
-            labels torch.Tensor (batch_size, sequence_length)
-    """
-    batch_text = []
-    for _ in range(batch_size):
-        batch_text.append(data[random.randint(0, len(data))]['text'])
-
-    # Tokenizer returns a dict { 'input_ids': list[], 'attention': list[] }
-    # but we need to convert to List [ dict ['input_ids': ..., 'attention': ... ]]
-    # annoying hack...
-    tokenized = tokenizer(batch_text)
-    tokenized = [dict(zip(tokenized,t)) for t in zip(*tokenized.values())]
-
-    # Produces the masked language model inputs aw dictionary dict {'inputs': tensor_batch, 'labels': tensor_batch}
-    # which can be used with the Bert Language model. 
-    collated_batch =  collator(tokenized)
-    return collated_batch['input_ids'], collated_batch['labels']
             
 def main(hparams):
     # Args
